Cleanup for Interop
