--- src/Common/src/Interop/Windows/Winsock/Interop.getsockopt.cs
@@ -1,57 +1,0 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System;
-using System.Runtime.InteropServices;
-using System.Net.Sockets;
-
-internal static partial class Interop
-{
-    internal static partial class Winsock
-    {
-        [DllImport(Interop.Libraries.Ws2_32, SetLastError = true)]
-        internal static extern SocketError getsockopt(
-                                           [In] SafeCloseSocket socketHandle,
-                                           [In] SocketOptionLevel optionLevel,
-                                           [In] SocketOptionName optionName,
-                                           [Out] out int optionValue,
-                                           [In, Out] ref int optionLength
-                                           );
-
-        [DllImport(Interop.Libraries.Ws2_32, SetLastError = true)]
-        internal static extern SocketError getsockopt(
-                                           [In] SafeCloseSocket socketHandle,
-                                           [In] SocketOptionLevel optionLevel,
-                                           [In] SocketOptionName optionName,
-                                           [Out] byte[] optionValue,
-                                           [In, Out] ref int optionLength
-                                           );
-
-        [DllImport(Interop.Libraries.Ws2_32, SetLastError = true)]
-        internal static extern SocketError getsockopt(
-                                           [In] SafeCloseSocket socketHandle,
-                                           [In] SocketOptionLevel optionLevel,
-                                           [In] SocketOptionName optionName,
-                                           [Out] out Linger optionValue,
-                                           [In, Out] ref int optionLength
-                                           );
-
-        [DllImport(Interop.Libraries.Ws2_32, SetLastError = true)]
-        internal static extern SocketError getsockopt(
-                                           [In] SafeCloseSocket socketHandle,
-                                           [In] SocketOptionLevel optionLevel,
-                                           [In] SocketOptionName optionName,
-                                           [Out] out IPMulticastRequest optionValue,
-                                           [In, Out] ref int optionLength
-                                           );
-
-        [DllImport(Interop.Libraries.Ws2_32, SetLastError = true)]
-        internal static extern SocketError getsockopt(
-                                           [In] SafeCloseSocket socketHandle,
-                                           [In] SocketOptionLevel optionLevel,
-                                           [In] SocketOptionName optionName,
-                                           [Out] out IPv6MulticastRequest optionValue,
-                                           [In, Out] ref int optionLength
-                                           );
-    }
-}
