+++ src/System.Net.Sockets.Legacy/tests/FunctionalTests/project.lock.json
@@ -631,7 +631,7 @@
           "lib/portable-net45+aspnetcore50+win+wpa81+wp80/xunit.core.dll": {}
         }
       },
-      "xunit.netcore.extensions/1.0.0-prerelease-00079": {
+      "xunit.netcore.extensions/1.0.0-prerelease-00080": {
         "dependencies": {
           "System.Diagnostics.Debug": "4.0.10",
           "System.IO": "4.0.10",
@@ -1999,12 +1999,12 @@
         "lib/portable-net45+aspnetcore50+win+wpa81+wp80/xunit.runner.utility.xml"
       ]
     },
-    "xunit.netcore.extensions/1.0.0-prerelease-00079": {
+    "xunit.netcore.extensions/1.0.0-prerelease-00080": {
       "serviceable": true,
-      "sha512": "YtqpdxTUJ776HVnP1JwpauW+ZXBNiLhoAFW2aG+hYyQDvZjrriakrMsFy/OA4aSYCCNCIU6de1iaoxtzLIbgqg==",
+      "sha512": "dHxf1OsXnrMolLeT2eqdyhi3wxQMwZG43jpmjpggXpyhLE3q7cSM7Chk8UWVx/JquqLSAWKg4yk+tPEQ7uz7bw==",
       "files": [
-        "xunit.netcore.extensions.1.0.0-prerelease-00079.nupkg",
-        "xunit.netcore.extensions.1.0.0-prerelease-00079.nupkg.sha512",
+        "xunit.netcore.extensions.1.0.0-prerelease-00080.nupkg",
+        "xunit.netcore.extensions.1.0.0-prerelease-00080.nupkg.sha512",
         "xunit.netcore.extensions.nuspec",
         "lib/dotnet/Xunit.NetCore.Extensions.dll"
       ]
